chore: start cleanup of vercel-ai provider
