+++ packages/vercel-ai/package.json
@@ -28,12 +28,14 @@
     "@ai-sdk/openai": "^1.3.10",
     "@voltagent/core": "^0.1.36",
     "ai": "^4.2.11",
+    "ts-pattern": "^5.7.1",
     "zod": "3.24.2"
   },
   "devDependencies": {
     "@types/jest": "^29.5.0",
     "@types/node": "^20.19.0",
     "@vitest/coverage-v8": "^3.2.3",
+    "@voltagent/internal": "workspace:^0.0.4",
     "eslint": "^8.0.0",
     "jest": "^29.5.0",
     "ts-jest": "^29.1.0",
