--- packages/vercel-ai/src/types.ts
@@ -1,6 +1,0 @@
-import type { LanguageModelV1CallOptions } from "ai";
-
-export type VercelProviderOptions = Omit<
-  LanguageModelV1CallOptions,
-  "messages" | "model" | "tools" | "maxSteps" | "schema"
->;
